Added a hook useAuth and moved AuthContext to the Contexts directory
